@@ -5,26 +5,25 @@ using Serilog;
 using System;
 using System.Linq;
 
-namespace Sample.Serilog.WebApi.Core.Extensions
+namespace Sample.Serilog.WebApi.Core.Extensions;
+
+public static class LogEnricherExtensions
 {
-    public static class LogEnricherExtensions
+    public static void EnrichFromRequest(IDiagnosticContext diagnosticContext, HttpContext httpContext)
     {
-        public static void EnrichFromRequest(IDiagnosticContext diagnosticContext, HttpContext httpContext)
-        {
-            diagnosticContext.Set("UserName", httpContext?.User?.Identity?.Name);
-            diagnosticContext.Set("ClientIP", httpContext.Connection.RemoteIpAddress.ToString());
-            diagnosticContext.Set("UserAgent", httpContext.Request.Headers["User-Agent"].FirstOrDefault());
-            diagnosticContext.Set("Resource", httpContext.GetMetricsCurrentResourceName());
-        }
+        diagnosticContext.Set("UserName", httpContext?.User?.Identity?.Name);
+        diagnosticContext.Set("ClientIP", httpContext.Connection.RemoteIpAddress.ToString());
+        diagnosticContext.Set("UserAgent", httpContext.Request.Headers["User-Agent"].FirstOrDefault());
+        diagnosticContext.Set("Resource", httpContext.GetMetricsCurrentResourceName());
+    }
 
-        public static string GetMetricsCurrentResourceName(this HttpContext httpContext)
-        {
-            if (httpContext == null)
-                throw new ArgumentNullException(nameof(httpContext));
+    public static string GetMetricsCurrentResourceName(this HttpContext httpContext)
+    {
+        if (httpContext == null)
+            throw new ArgumentNullException(nameof(httpContext));
 
-            Endpoint endpoint = httpContext.Features.Get<IEndpointFeature>()?.Endpoint;
+        Endpoint endpoint = httpContext.Features.Get<IEndpointFeature>()?.Endpoint;
 
-            return endpoint?.Metadata.GetMetadata<EndpointNameMetadata>()?.EndpointName;
-        }
+        return endpoint?.Metadata.GetMetadata<EndpointNameMetadata>()?.EndpointName;
     }
 }
