@@ -3,34 +3,33 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 
-namespace Sample.Serilog.WebApi.Core.HealthCheck
+namespace Sample.Serilog.WebApi.Core.HealthCheck;
+
+public class MyHealthCheck : IHealthCheck
 {
-    public class MyHealthCheck : IHealthCheck
+    private readonly IMyCustomService dependency;
+
+    public MyHealthCheck(IMyCustomService dependency)
     {
-        private readonly IMyCustomService dependency;
-
-        public MyHealthCheck(IMyCustomService dependency)
-        {
-            this.dependency = dependency;
-        }
-
-        public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context,CancellationToken cancellationToken = default(CancellationToken))
-        {
-            var result = dependency.IsHealthy() ? HealthCheckResult.Healthy() : HealthCheckResult.Unhealthy("Erro ao carregar as dependências.");
-            return Task.FromResult(result);
-        }
+        this.dependency = dependency;
     }
 
-    public interface IMyCustomService
+    public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context,CancellationToken cancellationToken = default(CancellationToken))
     {
-        public bool IsHealthy();
+        var result = dependency.IsHealthy() ? HealthCheckResult.Healthy() : HealthCheckResult.Unhealthy("Erro ao carregar as dependências.");
+        return Task.FromResult(result);
     }
+}
 
-    public class MyCustomService : IMyCustomService
+public interface IMyCustomService
+{
+    public bool IsHealthy();
+}
+
+public class MyCustomService : IMyCustomService
+{
+    public bool IsHealthy()
     {
-        public bool IsHealthy()
-        {
-            return new Random().NextDouble() > 0.6;
-        }
+        return new Random().NextDouble() > 0.6;
     }
 }
