@@ -1,48 +1,61 @@
 using System;
-using System.IO;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
+using HealthChecks.UI.Client;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Diagnostics.HealthChecks;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Sample.Serilog.WebApi.Core.Extensions;
+using Sample.Serilog.WebApi.Core.HealthCheck;
+using Sample.Serilog.WebApi.Core.Middleware;
 using Serilog;
 
-namespace Sample.Serilog.WebApi
+try
 {
-    public class Program
+    var builder = WebApplication.CreateBuilder(args);
+    builder.Host.AddSerilog(builder.Configuration, "API NoPaper");
+    Log.Information("Getting the motors running...");
+
+    builder.Services.AddRouting(options => options.LowercaseUrls = true);
+    builder.Services.AddSwaggerApi(builder.Configuration);
+    builder.Services.AddHeathCheckApi(builder.Configuration);
+    builder.Services.AddControllers();
+    builder.Services.AddScoped<IMyCustomService, MyCustomService>();
+
+    var app = builder.Build();
+
+    if (app.Environment.IsDevelopment())
     {
-        public static IConfiguration Configuration { get; } = new ConfigurationBuilder()
-            .SetBasePath(Directory.GetCurrentDirectory())
-            .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-            .AddJsonFile($"appsettings.{Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") ?? "Production"}.json", optional: true)
-            .AddEnvironmentVariables()
-            .Build();
-
-        public static void Main(string[] args)
-        {
-            SerilogExtensions.AddSerilogApi(Configuration);
-
-            try
-            {
-                Log.Information("Iniciando a aplica��o...");
-
-                CreateHostBuilder(args).Build().Run();
-            }
-            catch (Exception ex)
-            {
-                Log.Fatal(ex, "Erro lan�ado pela aplica��o.");
-            }
-            finally
-            {
-                Log.CloseAndFlush();
-            }
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .UseSerilog()
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
+        app.UseDeveloperExceptionPage();
     }
+
+    app.UseSerilogRequestLogging(opts => opts.EnrichDiagnosticContext = LogEnricherExtensions.EnrichFromRequest);
+
+    app.UseMiddleware<RequestSerilLogMiddleware>();
+    app.UseMiddleware<ErrorHandlingMiddleware>();
+
+    app.UseSwaggerDocApi();
+    app.UseHttpsRedirection();
+    app.UseRouting();
+    app.UseHealthCheckApi();
+    app.UseEndpoints(endpoints =>
+    {
+        endpoints.MapControllers();
+        endpoints.MapHealthChecks("/hc",
+            new HealthCheckOptions
+            {
+                Predicate = _ => true,
+                ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse,
+            });
+    });
+
+    app.Run();
+}
+catch (Exception ex)
+{
+    Log.Fatal(ex, "Host terminated unexpectedly");
+}
+finally
+{
+    Log.Information("Server Shutting down...");
+    Log.CloseAndFlush();
 }
